Update remaining usages of LOCALHOST_HOSTNAME
