+++ localstack/services/stepfunctions/asl/component/state/state_execution/state_task/service/state_task_service_api_gateway.py
@@ -15,7 +15,6 @@ from localstack.aws.api.stepfunctions import HistoryEventType, TaskFailedEventDe
 from localstack.constants import (
     APPLICATION_JSON,
     HEADER_CONTENT_TYPE,
-    LOCALHOST_HOSTNAME,
     PATH_USER_REQUEST,
 )
 from localstack.services.stepfunctions.asl.component.common.error_name.custom_error_name import (
@@ -34,6 +33,7 @@ from localstack.services.stepfunctions.asl.eval.environment import Environment
 from localstack.services.stepfunctions.asl.eval.event.event_detail import EventDetails
 from localstack.utils.collections import select_from_typed_dict
 from localstack.utils.strings import long_uid
+from localstack.utils.urls import localstack_host
 
 LOG = logging.getLogger(__name__)
 
@@ -164,7 +164,7 @@ class StateTaskServiceApiGateway(StateTaskServiceCallback):
         #  there's an argument to be made that this may mast implementation mistakes: investigate further.
         url_spec = urlparse(api_endpoint)
         url_path = url_spec.path
-        if not url_path.endswith(LOCALHOST_HOSTNAME):
+        if not url_path.endswith(localstack_host().host):
             return api_endpoint
         path_parts = url_path.split(".")
         api_id = path_parts[0]
